chore: move Page/PageParams outside of export
as PageParams is used in deduplication as well and Page should be
together with PageParams
