+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/notification/ProgramNotificationTemplateController.java
@@ -88,7 +88,7 @@ public class ProgramNotificationTemplateController
       long total = programNotificationTemplateService.countProgramNotificationTemplates(params);
       return Page.withPager(
           ProgramNotificationTemplateSchemaDescriptor.PLURAL,
-          org.hisp.dhis.tracker.export.Page.withTotals(
+          org.hisp.dhis.tracker.Page.withTotals(
               instances, params.getPage(), params.getPageSize(), total));
     }
